lib/config, gui: Deprecate the config entries for pending stuff.
Rename the config structure members to DeprecatedPendingDevices and
DeprecatedPendingFolders.  Remove all code referencing the old names.

These entries are regarded as non-vital information which will come
back after being cleaned from the config if still relevant.  The
notifications NOT appearing again means that either the associated
device is not connecting to us or it does not offer the previously
pending thing anymore.  In both cases, keeping the pending
notification would be of no advantage because even if accepted,
syncing the offered thing would require a successful connection
regardless.

Remove empty initialization of the pendingFolders member in the GUI's
addDevice() function as well.