lib/model: Avoid trying to delete non-existant DB PendingFolders.
Trying to remove entries for all shared device IDs of all known
folders after a configuration change causes many database lookups.
For newly added folders, this is unavoidable as there may be several
obsolete entries (now shared) and also remaining valid entries (now
requesting sharing of a locally known folder).  A further optimization
here might be to iterate over the corresponding PendingFolder entries
and looking up in a map of shared device IDs instead of trying to
delete each in turn.  That is not done with this change.

For an already known folder, build a drop-list of device IDs by
comparing with the previous configuration to see which devices it was
additionally shared with.  That still might include folder / device
combinations for which no PendingFolder entry even existed, but is
already slightly more efficient.