lib/db: Add helper type for Pending* cleanup functions.
Introduce the named type DropListObserved for a nested map structure.
First level is the device ID, including those for which PendingFolders
are allowed at all.  Second level is the folder ID with a boolean
value, which marks those folders which should not be pending.

Add helper methods to populate this type.