lib/db: Add CleanPendingDevices() with iterating semantics.
Just like CleanPendingFolders(), this iterates over all PendingDevice
entries once and decides whether to remove them based on a
pre-assembled lookup map.  Reuse the same DropListObserved nested map
type.  This implements the semantics outlined for the custom type, for
devices.

Note that this is work-in-progress and does not build without changes
to the calling code.