lib/model: Defer pending stuff cleanup to single iteration passes.
Instead of calling RemovePendingDevice() and RemovePendingFolder() in
many places, assemble a data structure marking the actions to be taken
for each.  When all decisions are collected, call the CleanPending*()
functions to iterate over all entries once and apply the cached
decisions.

As Pending... entries in the DB are usually few, this will be more
efficient.  It also avoids repeated DB record lookups for
deletion (which might not even exist), instead only doing (fast)
lookups on the in-memory nested map type.