lib/model: Broaden cleanup of PendingFolder for removed folders.
As discussed in the code comment, ALL PendingFolder references should
be dropped for a completely removed folder, not just the ones it was
previously shared with (which technically could not have been pending
anymore).

Include the complete list of known devices for this folder's cleanup
decisions.  Since that is a positive list, that's all we can do.  Even
more (unknown) devices will not be listed at all in the decision map
and therefore cause associated PendingFolders to be removed as well.