lib/db: Fix inverted logic in CleanPendingFolders().
If no folders are listed for a device in the drop list, that means
none should be removed.  Fix the check for a nil value leading to
deletion, when it should actually just skip looking up the folder ID
and continue.

Also use len(map) for robustness which handles empty maps and nil the
same.