lib/db: Remove unneeded myID parameter from MarkFolder().
In the (impossible) case that there are Pending... records for our own
device ID, those should definitely be removed.  Such entries in the
drop-list come from the fact that FolderConfiguration always includes
the local device ID association as well.  For CleanPendingFolders(),
that's perfectly fine and does not need to be avoided.

For CleanPendingDevices(), it would actually prevent removing such
unwanted DB records.  To solve it for that iteration, just remove the
local device ID from the map right before the clean-up.

Note that this is not easily possible in onServe() because the config
wrapper has no API to access MyID.

Consolidate some comments about the collection in forgetPending.