lib/model: Make sure to drop pending entries for removed devices.
As another precautious measure, include the IDs of any devices which
are just being removed in the forgetPending decision matrix before
passing to CleanPendingDevices().  There *should not* be any
PendingDevice entries for these since they were previously already
added.  However it's rather cheap to mark them correctly in the
drop-list.  And just in case one did slip through the cracks, this
prevents from showing a GUI notification for a device that the user
just deliberately removed.