lib/db: Extend and clean up log formatting.
When dropping an invalid DB entry, include the key content in the log
message as hexadecimal bytes.  That's the last chance to capture the
value if needed for debug analysis.

Consistently print the whole key.  Especially the value returned when
DeviceFromPendingFolderKey() fails is always nil and therefore not
useful.

Use %x and %s format conversions consistently.