lib/api: Serialize pending devices and folders as maps, not arrays.
Move the id and deviceID fields out of the enumerated objects and use
them as attribute names in a dictionary mapping (JSON object)
instead.  This applies to the GET /rest/cluster/pendingdevices
enumeration, the GET /rest/cluster/pendingfolders enumeration and the
offeredBy attributes within the latter.

Give protocol.DeviceID.MarshalText() a non-pointer receiver argument
to make automatic marshalling work when DeviceIDs are involved as keys
of a map.

Remove the toJson...Slice() methods and associated helper types,
instead serializing the maps themselves.  Only exception is
toJsonPendingFolderMap(), which wraps the offeredBy attribute name
around the offering device map.  This greatly simplifies the
serialization code and to some extent also the usage within the GUI.

Adjust the API usage in the GUI to store the pendingFolders and
pendingDevices as objects instead of arrays.  This affects the way
pending... entries are converted to ignored... ones since the former
now need the ID as a separate piece of information while the latter
include it as an attribute.