lib/config: Apply defaults to folders/devices on config load (ref #6748)
This makes it so that `default:"..."` struct tags on devices and folders
actually take effect during JSON/XML loading, like it does for general
and GUI options.

This means we can add a field to a device/folder and have it take its
default value on config load without writing a config migration. (I
suspect we've missed this in the past because the current behavior is
unintuitive...)