build, lib/upgrade: Compare build metadata after plus in version (ref #6758)
With this change we take into account the stuff after "+" in version
comparisons, when all else is equal. Having something after the plus is
newer than not having something. If both versions have something after
the plus they are compared in the same manner as prerelease parts (after
"-") are compared: in dot separated components, numerically if both are
numbers or alphabetically if both are non-numbers. Having differing
types in the same position in the version string is confusing, don't do
that.

Also changes the build metadata to be "+commits.hash" instead of
"+commits-hash" in order to get a dot separator so the number of commits
gets compared numerically.