lib/api, lib/model: Improve folder completion API (fixes #6075)
This makes the /rest/db/completion call more useful and user friendly.

Prior to this change the only supported use was with both "folder" and
"device" parameters, where the device is a remote device. This then
gives the completion status for specifically that folder on that device.

With this change we also allow the following:

- Omitting the "device" parameter, in which case it means the local
  device (and returns the local completion)

- Explicitly specifying the local device ID, which means the same thing.

- Omitting the "folder" parameter, in which case it means all folders.

- Omitting both, which logically then means the completion percentage
  for all folders as a whole on this device.

This should be closer to what people expect usage wise, while also now
giving a single place to check for a completion percentage for the local
device or any remote device as a whole.