gui: Reuse cached probing results for same address on different ID.
Simplify the remoteGUICache cleanup by just rebuilding it every time,
which is presumably easier on the JavaScript engine than deleting an
object property.  The previous approach was taken to safeguard against
modifying the object within the probing response callback while it is
being replaced in refreshRemoteGUI().  But apparently it's safe to do
because JavaScript is single-threaded and AngularJS won't start the
probing requests until refreshRemoteGUI() has finished.

Make sure to copy cached result values to the idToRemoteGUI mapping
for every device ID which has the same address configured.  Otherwise
the response callback would only apply the result to one specific ID.