gui: Display remote GUI port in title tooltip, not the URL.
The URL is already displayed in cleartext and by the the browser when
hovering on the link.  Another repetition within the table cell's
title tooltip doesn't really help.

On the other hand, showing the configured port even in the case when
probing failed and thus the URL is empty, seeing the port on hovering
might give a clue for why it failed.