gui: Properly encode IPv6 URLs containing a zone identifier.
According to RFC6874, the % delimiter character for the zone
identifier of a link-local address (see RFC4007) must be
percent-encoded within URIs.  Apply this encoding when generating a
remote GUI link target and for the URLs being probed.

That still causes trouble in modern browsers, because they don't
recognise the zone identifier symtax in IPv6 literals (with or without
percent-encoding), or do not support link-local addresses in HTML
links (or the address input field) at all.  But that is their bug,
Syncthing now generates an RFC-compliant URL.