lib/model: Store candidate devices from ClusterConfig to database.
When iterating over a ClusterConfig message, currently only the remote
side's view of our own device and of themselves is considered.  Extend
those checks for devices that would be candidates for additional
sharing links within the cluster, since they share the same folder ID,
but not directly with our local device.  Those might be completely
unknown devices so far.

The new internal method ccHandleFolderCandidates() encapsulates that
functionality.  Errors during DB saving are just logged as a warning
and otherwise ignored.  Additional details like the introducer's
suggested name for the candidate device are only stored if it wasn't
already known locally.

Add some FIXME comments for branches that skip to the next folder
before we had a chance to record the candidate devices.