lib/model, lib/db: Forget candidate links no longer in ClusterConfig.
Implement a convenience function RemoveCandidateLinksBeforeTime() to
remove all entries for a given introducer that are older than the
specified timestamp.  It reuses readCandidateLink() with the side
effect of dropping invalid DB entries.  Returns a list of removed
valid CandidateLink entries for consistency, which is not actually
used right now as push notifications (events) for changes in collected
candidates are not required.

Hook that up to the ClusterConfig handling to discard older entries
based on the timestamp when receiving the message.