lib/db: Provide dummy data for testing candidates feature.
Only for development purposes, can be deleted later or reused in
tests.

Includes replacement functions to return made-up sample data instead
of the real functions for enumerating the DB entries.

CandidateLinksDummyData() can be used to populate an actual DB with
the sample data which the API endpoint documentation is based on.