lib/api: Returns tags in version as list
The GUI expected the version.tags field to be a list of current build
tags, but the thing the API sent was actually the comma separated string
set by the build script. (This was broken in d53a2567a.) First of all
this is the wrong format, secondly it doesn't include all the things we
actually consider tags. This fixes that.

Technically it's a breaking API change. I don't think anyone cares; it's
a bugfix of the previous unintentional API change.