gui: Fix missing sharing device when adding pending folder.
When adding a folder from the pending folder from another device
through the Add button on the "New Folder" notification panel, the
offering device should be preselected for sharing with.  The selection
works, but the checkmark for that device is not shown at all.

Do not filter out the preselected device ID when building the
currentSharing.unrelated member in addFolderAndShare() to actually
show all devices under the "Unshared Devices" heading.