lib/model: Stricter filtering of candidate links.
Do not consider candidate suggestions from an introducer with which we
don't actually share the local folder.  Just because it offers the
folder to us and another peer, doesn't mean the other peer already has
access to the data of our local folder with this ID.

Also avoid adding candidate links for locally known third party
devices for which the same folder ID is configured to be ignored.  If
we deliberately reject offers for this folder from that remote, we
should not suggest sharing to it because of an introducer's
ClusterConfig.