lib/protocol: Hide repeated data blocks in a given file
Adds the block offset as additional data into the AES-SIV encryption of
the block hash, thus making it not possible to correlate blocks at
different offsets within a file.

Also, decrypts the block hash and passes it on to the next layer so that
it can verify the data as read.