gui: Only check candidates when editing existing folders.
After the last merge, querying the candidates API endpoint happened in
editFolder().  That function is also used for editing the default
folder config, where candidates do not make sense.  Because it called
initShareEditing() and editFolderModal() in one go, rename to
initPathEditing() and split up so it only does the folder path
manipulation.  That allows moving the candidates query to
editFolderExisting().