lib/model: Don't remove non-empty .stfolder
When a folder is removed we remove the marker. But there may be things
stored in it that have usefulness past the lifetime of the folder in
Syncthing, in particularly the folder ID marker for encrypted folders.

This changes the behavior to remove empty markers (or if the marker's a
file...) but otherwise rename it to `.stfolder.removed-$timestamp` so
that it's out of the way but the contents aren't lost.