gui: Move the QR code button next to device ID in editDevice modal.
The button shows a different representation of the device ID input
field contents, so move it right next to that one.  There are two
incarnations because in the editingExisting case, the read-only device
ID is not actually an HTML input.  Treat both as a Bootstrap input
group with attached button.

Note this removes access to the "Show QR" function when any other tab
than "General" is active.  That is intentional to better guide the
user about which device ID will be shown (not the local one).