lib/api, gui: Enumerate non-failed discovery methods as well.
For least API breakage, just include them with a null value in the
discoveryErrors object.  That effectively deprecates the
discoveryMethods field, which now always matches the number of entries
in discoveryErrors.  API users just need to adapt to checking whether
the error is non-null.

For consistency with the connectionServiceStatus field, it should be
renamed to discoveryStatus and mirror the structure with subordinate
objects, but there is no real benefit that justifies this API
breakage.