lib/fs: Add encoder filesystem (fixes #1734)
A NewEncoderFilesystem ensures that paths that
contain characters that are reserved on
filesystems such as NTFS can be safety stored.
It does this by replacing the reserved
characters with UNICODE characters in
the private use area (\uf000-\uf07f). This
conversion is compatible with Cygwin, Git-Bash,
Msys2, Windows Subsystem for Linux (WSL), and
other platforms.

For reference, see:
https://cygwin.com/cygwin-ug-net/using-specialnames.html
http://msdn.microsoft.com/en-us/library/aa365247%28VS.85%29.aspx
https://en.wikipedia.org/wiki/Filename#In_Windows
https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file

For implementations, see
https://github.com/mirror/newlib-cygwin/blob/fb01286fab9b370c86323f84a46285cfbebfe4ff/winsup/cygwin/path.cc#L435
https://github.com/billziss-gh/winfsp/blob/6e3a8f70b2bd958960012447544d492fc6a2f1af/src/shared/ku/posix.c#L1250

Signed-off-by: Ross Smith II <ross@smithii.com>