gui: Use large modals for Listeners and Discovery
Right now, small modals are used for Listeners and Discovery. However,
especially in the case Discovery, quite a lot of information is crammed
into the modal, which basically fills it completely in order to fit the
long sentences and long error messages. Because of this, change the type
of the modal to "large", so that everything has more space to fit more
comfortably.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>