gui: Display identicons for discovered device IDs.
In the Add Device modal dialog, there is a list of up to five locally
discovered device IDs.  Choosing from them is a great time-saver, but
identifying which is the wanted one requires comparing (partial)
device ID strings.  Especially when integrating a new device into an
existing cluster, one can more easily remember the identicon shown on
one of the existing cluster devices.  Make it easy to recognize that
device in the discovered ID list by replacing the bullet points with
the corresponding identicon.