cmd/syncthing: Simplify logging for --generate option.
Using the app-wide debug logger mechanism in generate() is not very
useful, as the --logfile argument is interpreted by the monitor
process and that cannot be in use when generate() is called.
Switching to basic log.Println() calls allows to factor out this
function into its own package for a subcommand, which won't need its
own logger instance.

Visible change: No more timestamp prefixes.