cmd/syncthing: Accept new GUI credentials in generate subcommand.
Add the options --gui-user and --gui-password to "generate", but not
the "serve --generate" option form.  If either is given, an existing
config will not abort the command, but rather load, modify and save it
with the new credentials.  The password can be read from standard
input by passing only a single dash as argument.

Config modification is skipped if the value matches what's already in
the config.  Especially passing in the hashed password will not cause
a modification, but giving the same (plaintext) password again will
result in varying hashes.